+++ src/PLand.cc
@@ -1,6 +1,7 @@
 #include "pland/PLand.h"
 #include "fmt/core.h"
 #include "ll/api/data/KeyValueDB.h"
+#include "ll/api/i18n/I18n.h"
 #include "mc/world/actor/player/Player.h"
 #include "mc/world/level/BlockPos.h"
 #include "mod/MyMod.h"
@@ -17,11 +18,16 @@
 #include <mutex>
 #include <shared_mutex>
 #include <string>
+#include <unordered_set>
 #include <utility>
 #include <vector>
 
 
 namespace land {
+std::string PlayerSettings::SYSTEM_LOCALE_CODE() { return "system"; }
+std::string PlayerSettings::SERVER_LOCALE_CODE() { return "server"; }
+
+
 void PLand::_loadOperators() {
     if (!mDB->has(DB_KEY_OPERATORS())) {
         mDB->set(DB_KEY_OPERATORS(), "[]"); // empty array
@@ -35,7 +41,15 @@ void PLand::_loadPlayerSettings() {
         mDB->set(DB_KEY_PLAYER_SETTINGS(), "{}"); // empty object
     }
     auto settings = JSON::parse(*mDB->get(DB_KEY_PLAYER_SETTINGS()));
-    JSON::jsonToStructNoMerge(settings, mPlayerSettings);
+    if (!settings.is_object()) {
+        throw std::runtime_error("player settings is not an object");
+    }
+
+    for (auto& [key, value] : settings.items()) {
+        PlayerSettings settings_;
+        JSON::jsonToStructTryPatch(value, settings_);
+        mPlayerSettings.emplace(key, std::move(settings_));
+    }
 }
 
 void PLand::_loadLands() {
@@ -73,9 +87,7 @@ void PLand::_initLandMap() {
             auto  chunkID      = PLand::EncodeChunkID(ch.x, ch.z);
             auto& chunkLandVec = chunkMap[chunkID]; // 区块领地数组
 
-            if (!some(chunkLandVec, land->getLandID())) {
-                chunkLandVec.push_back(land->getLandID());
-            }
+            chunkLandVec.insert(land->getLandID());
         }
     }
     my_mod::MyMod::getInstance().getSelf().getLogger().info("初始化领地缓存系统完成");
@@ -86,16 +98,10 @@ void PLand::_updateLandMap(LandData_sptr const& ptr, bool add) {
     for (auto& ch : chunks) {
         auto& chunkLands = mLandMap[ptr->mLandDimid][EncodeChunkID(ch.x, ch.z)];
 
-        auto iter = std::find(chunkLands.begin(), chunkLands.end(), ptr->mLandID);
-
         if (add) {
-            if (iter == chunkLands.end()) {
-                chunkLands.push_back(ptr->mLandID);
-            }
+            chunkLands.insert(ptr->getLandID());
         } else {
-            if (iter != chunkLands.end()) {
-                chunkLands.erase(iter);
-            }
+            chunkLands.erase(ptr->getLandID());
         }
     }
 }
@@ -103,6 +109,9 @@ void PLand::_refreshLandRange(LandData_sptr const& ptr) {
     _updateLandMap(ptr, false);
     _updateLandMap(ptr, true);
 }
+
+LandID PLand::getNextLandID() { return mNextLandID++; }
+
 } // namespace land
 
 
@@ -186,6 +195,10 @@ bool PLand::removeOperator(UUIDs const& uuid) {
     mLandOperators.erase(iter);
     return true;
 }
+std::vector<UUIDs> const& PLand::getOperators() const {
+    std::shared_lock<std::shared_mutex> lock(mMutex);
+    return mLandOperators;
+}
 
 
 PlayerSettings* PLand::getPlayerSettings(UUIDs const& uuid) {
@@ -216,10 +229,10 @@ bool PLand::addLand(LandData_sptr land) {
         return false;
     }
 
-    LandID id = generateLandID();
+    LandID id = getNextLandID();
     if (hasLand(id)) {
         for (size_t i = 0; i < 3; i++) {
-            id = generateLandID();
+            id = getNextLandID();
             if (!hasLand(id)) {
                 break;
             }
@@ -290,6 +303,17 @@ std::vector<LandData_sptr> PLand::getLands() const {
     }
     return lands;
 }
+std::vector<LandData_sptr> PLand::getLands(std::vector<LandID> const& ids) const {
+    std::shared_lock<std::shared_mutex> lock(mMutex);
+
+    std::vector<LandData_sptr> lands;
+    for (auto id : ids) {
+        if (auto iter = mLandCache.find(id); iter != mLandCache.end()) {
+            lands.push_back(iter->second);
+        }
+    }
+    return lands;
+}
 std::vector<LandData_sptr> PLand::getLands(LandDimid dimid) const {
     std::shared_lock<std::shared_mutex> lock(mMutex);
 
@@ -336,10 +360,11 @@ LandPermType PLand::getPermType(UUIDs const& uuid, LandID id, bool ignoreOperato
     return LandPermType::Guest;
 }
 
-LandID PLand::generateLandID() { return mNextLandID++; }
 
 LandData_sptr PLand::getLandAt(BlockPos const& pos, LandDimid dimid) const {
     std::shared_lock<std::shared_mutex> lock(mMutex);
+
+    std::unordered_set<LandData_sptr> result;
 
     ChunkID chunkId = EncodeChunkID(pos.x >> 4, pos.z >> 4);
     auto    dimIt   = mLandMap.find(dimid); // 查找维度
@@ -350,17 +375,39 @@ LandData_sptr PLand::getLandAt(BlockPos const& pos, LandDimid dimid) const {
                 auto landIt = mLandCache.find(landId); // 查找领地
                 if (landIt != mLandCache.end()
                     && landIt->second->getLandPos().hasPos(pos, !landIt->second->is3DLand())) {
-                    return landIt->second;
+                    // return landIt->second;
+                    result.insert(landIt->second);
                 }
             }
         }
     }
+
+    if (!result.empty()) {
+        if (result.size() == 1) {
+            return *result.begin();
+        }
+
+        // 重建领地关系
+        for (auto& land : result) {
+            if (land->isSubLand()) {
+                return land;
+            }
+        }
+    }
+
     return nullptr;
 }
-std::vector<LandData_sptr> PLand::getLandAt(BlockPos const& center, int radius, LandDimid dimid) const {
+std::unordered_set<LandData_sptr> PLand::getLandAt(BlockPos const& center, int radius, LandDimid dimid) const {
     std::shared_lock<std::shared_mutex> lock(mMutex);
 
-    std::vector<LandData_sptr> lands;
+    auto dimIter = mLandMap.find(dimid); // 查找维度
+    if (dimIter == mLandMap.end()) {
+        return {};
+    }
+
+    auto&                             dim = dimIter->second;
+    std::unordered_set<ChunkID>       visitedChunks; // 记录已访问的区块
+    std::unordered_set<LandData_sptr> lands;
 
     int minChunkX = (center.x - radius) >> 4;
     int minChunkZ = (center.z - radius) >> 4;
@@ -370,15 +417,17 @@ std::vector<LandData_sptr> PLand::getLandAt(BlockPos const& center, int radius, 
     for (int x = minChunkX; x <= maxChunkX; ++x) {
         for (int z = minChunkZ; z <= maxChunkZ; ++z) {
             ChunkID chunkId = EncodeChunkID(x, z);
-            auto    dimIt   = mLandMap.find(dimid); // 查找维度
-            if (dimIt != mLandMap.end()) {
-                auto chunkIt = dimIt->second.find(chunkId); // 查找区块
-                if (chunkIt != dimIt->second.end()) {
-                    for (const auto& landId : chunkIt->second) {
-                        auto landIt = mLandCache.find(landId); // 查找领地
-                        if (landIt != mLandCache.end() && landIt->second->isRadiusInLand(center, radius)) {
-                            lands.push_back(landIt->second);
-                        }
+            if (visitedChunks.find(chunkId) != visitedChunks.end()) {
+                continue; // 如果区块已经访问过，则跳过
+            }
+            visitedChunks.insert(chunkId);
+
+            auto chunkIt = dim.find(chunkId); // 查找区块
+            if (chunkIt != dim.end()) {
+                for (const auto& landId : chunkIt->second) {
+                    auto landIt = mLandCache.find(landId); // 查找领地
+                    if (landIt != mLandCache.end() && landIt->second->isRadiusInLand(center, radius)) {
+                        lands.insert(landIt->second);
                     }
                 }
             }
@@ -386,10 +435,17 @@ std::vector<LandData_sptr> PLand::getLandAt(BlockPos const& center, int radius, 
     }
     return lands;
 }
-std::vector<LandData_sptr> PLand::getLandAt(BlockPos const& pos1, BlockPos const& pos2, LandDimid dimid) const {
+std::unordered_set<LandData_sptr> PLand::getLandAt(BlockPos const& pos1, BlockPos const& pos2, LandDimid dimid) const {
     std::shared_lock<std::shared_mutex> lock(mMutex);
 
-    std::vector<LandData_sptr> lands;
+    auto dimIter = mLandMap.find(dimid); // 查找维度
+    if (dimIter == mLandMap.end()) {
+        return {};
+    }
+
+    auto&                             dim = dimIter->second;
+    std::unordered_set<ChunkID>       visitedChunks;
+    std::unordered_set<LandData_sptr> lands;
 
     int minChunkX = std::min(pos1.x, pos2.x) >> 4;
     int minChunkZ = std::min(pos1.z, pos2.z) >> 4;
@@ -399,15 +455,17 @@ std::vector<LandData_sptr> PLand::getLandAt(BlockPos const& pos1, BlockPos const
     for (int x = minChunkX; x <= maxChunkX; ++x) {
         for (int z = minChunkZ; z <= maxChunkZ; ++z) {
             ChunkID chunkId = EncodeChunkID(x, z);
-            auto    dimIt   = mLandMap.find(dimid); // 查找维度
-            if (dimIt != mLandMap.end()) {
-                auto chunkIt = dimIt->second.find(chunkId); // 查找区块
-                if (chunkIt != dimIt->second.end()) {
-                    for (const auto& landId : chunkIt->second) {
-                        auto landIt = mLandCache.find(landId); // 查找领地
-                        if (landIt != mLandCache.end() && landIt->second->isAABBInLand(pos1, pos2)) {
-                            lands.push_back(landIt->second);
-                        }
+            if (visitedChunks.find(chunkId) != visitedChunks.end()) {
+                continue;
+            }
+            visitedChunks.insert(chunkId);
+
+            auto chunkIt = dim.find(chunkId); // 查找区块
+            if (chunkIt != dim.end()) {
+                for (const auto& landId : chunkIt->second) {
+                    auto landIt = mLandCache.find(landId); // 查找领地
+                    if (landIt != mLandCache.end() && landIt->second->isAABBInLand(pos1, pos2)) {
+                        lands.insert(landIt->second);
                     }
                 }
             }
@@ -429,6 +487,8 @@ ChunkID PLand::EncodeChunkID(int x, int z) {
     if (x >= 0) signBits |= (1ULL << 63);
     if (z >= 0) signBits |= (1ULL << 62);
     return signBits | (ux << 31) | (uz & 0x7FFFFFFF);
+    // Memory layout:
+    // [signBits][x][z] (signBits: 2 bits, x: 31 bits, z: 31 bits)
 }
 std::pair<int, int> PLand::DecodeChunkID(ChunkID id) {
     bool xPositive = (id & (1ULL << 63)) != 0;