+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0]
+
+### Added
+
+- 新增 `en_US`、`ru_RU` 语言文件 #17
+- 新增 `/pland list op` 命令 #15
+- 新增 `/pland set language` 命令 #19
+- 新增 `/pland this` 命令
+- `Config` 添加 `subLand` 子领地相配置 #18
+
+### Changed
+
+- 移除 `Particle`、`LandDraw`
+- 优化 `PLand::getLandAt` 查询
+- 重构 `Calculate` 为 `PriceCalculate`
+- 重构部分代码
+
 ## [0.6.0] - 2025-03-01
 
 ### Changed